Reorganized files for clarity
