+++ frontend/src/pages/ProductPage.tsx
@@ -1,20 +1,20 @@
-import * as React from 'react';
-import TestBut from '../components/TestBut';
-import Stack from '@mui/joy/Stack';
-import Box from '@mui/joy/Box';
-import ProductCard from '../components/ProductCard';
-import Pagination from '@mui/material/Pagination';
-import PaginationItem from '@mui/material/PaginationItem';
-import ArrowBackIcon from '@mui/icons-material/ArrowBack';
-import ArrowForwardIcon from '@mui/icons-material/ArrowForward';
-import Grid from '@mui/material/Grid'; 
-import '../assets/christinasStyleSheet.css';
+import * as React from "react";
+import NavigationBar from "../components/NavigationBar";
+import Stack from "@mui/joy/Stack";
+import Box from "@mui/joy/Box";
+import ProductCard from "../components/ProductCard";
+import Pagination from "@mui/material/Pagination";
+import PaginationItem from "@mui/material/PaginationItem";
+import ArrowBackIcon from "@mui/icons-material/ArrowBack";
+import ArrowForwardIcon from "@mui/icons-material/ArrowForward";
+import Grid from "@mui/material/Grid";
+import "../assets/Custom-Style-Sheets/christinasStyleSheet.css";
 
 export default function ProductsPage() {
   return (
     <div>
-    <TestBut/>
-          <div
+      <NavigationBar />
+      <div
         style={{
           backgroundRepeat: "no-repeat",
           backgroundSize: "cover",
@@ -28,29 +28,48 @@ export default function ProductsPage() {
           zIndex: "-1",
         }}
       ></div>
-    <Box sx={{flexGrow:1, mt:"4rem", ml:"10rem"}}>
-        <Grid container spacing={4} >
-            <Grid item lg={3}><ProductCard/></Grid>
-            <Grid item lg={3}><ProductCard/></Grid>
-            <Grid item lg={3}><ProductCard/></Grid>
-            <Grid item lg={3}><ProductCard/></Grid>
-            <Grid item lg={3}><ProductCard/></Grid>
-            <Grid item lg={3}><ProductCard/></Grid>
-            <Grid item lg={3}><ProductCard/></Grid>
-            <Grid item lg={3}><ProductCard/></Grid>
-            </Grid>
-    </Box>
-          <Stack spacing={2} justifySelf="center" sx={{mt:"5rem"}}>
+      <Box sx={{ flexGrow: 1, mt: "4rem", ml: "10rem" }}>
+        <Grid container spacing={4}>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+          <Grid item lg={3}>
+            <ProductCard />
+          </Grid>
+        </Grid>
+      </Box>
+      <Stack spacing={2} justifySelf="center" sx={{ mt: "5rem" }}>
         <Pagination
-        count={10}
-        renderItem={(item) => (
+          count={10}
+          renderItem={(item) => (
             <PaginationItem
-            slots={{ previous: ArrowBackIcon, next: ArrowForwardIcon }}
-            {...item} sx={{backgroundColor:"white"}}
+              slots={{ previous: ArrowBackIcon, next: ArrowForwardIcon }}
+              {...item}
+              sx={{ backgroundColor: "white" }}
             />
-        )} variant="outlined" shape="rounded" 
+          )}
+          variant="outlined"
+          shape="rounded"
         />
-    </Stack>
+      </Stack>
     </div>
   );
 }
