--- frontend/src/pages/TestPage.tsx
@@ -1,29 +1,0 @@
-
-import TestUI from '../components/TestUI';
-import TestBut from '../components/TestBut';
-
-const TestPage = () => {
-    return (
-        <div>
-            <TestBut/>
-            <div
-                style={{
-                    backgroundImage: `url('/frame.svg')`, 
-                    backgroundRepeat: 'no-repeat',
-                    backgroundSize: 'cover',
-                    backgroundPosition: 'center',
-                    width: '100%',
-                    height: '100vh', 
-                    position: 'absolute',
-                    top: '0',
-                    left: '0',
-                    zIndex: '-1', 
-                }}
-
-                
-            ></div>
-            <TestUI />
-        </div>
-    );
-}
-export default TestPage;
